Added MyRecipes and child components
