+++ client/src/App.js
@@ -3,7 +3,10 @@ import { Route, Routes } from "react-router-dom";
 import Home from "./components/home/index";
 import CreateRecipe from "./components/recipes/create";
 import ShowRecipes from "./components/recipes/index";
+import MyRecipes from "./components/recipes/myRecipes";
 import ShowIngredients from "./components/ingredients/index";
+
+
 import CreateIngredients from "./components/ingredients/create";
 import Logon from "./components/auth/create";
 import Register from "./components/auth/register";
@@ -19,7 +22,7 @@ const App = () => {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/recipe/create" element={<CreateRecipe />} />
-        <Route path="/recipe/" element={<ShowRecipes />} />
+        <Route path="/recipe/" element={<MyRecipes />} />
         <Route path="/ingredient/" element={<ShowIngredients />} />
         <Route path="/ingredient/create" element={<CreateIngredients />} />
         <Route path="/auth/" element={<Logon />} />
